janosik-ea/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
MetaTrader5>=5.0.30

# ML & RL
stable-baselines3>=1.8.0
gymnasium>=0.27.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Backtesting & Analysis
ta>=0.10.2
plotly>=5.13.0

# Monitoring
python-telegram-bot>=20.0
dash>=2.9.0
plotly-dash>=2.9.0

# Utils
python-dateutil>=2.8.2
requests>=2.28.0
